+++ src/routes/user_route.py
@@ -3,10 +3,12 @@ from sqlalchemy.orm import Session
 
 from dependencies import get_current_active_user, get_db
 from domain.schemas.bookrequest_schemas import DomainReqGetBookRequest
+from domain.services.book_review_service import service_read_reviews_by_user_id
 from domain.services.bookrequest_service import service_read_bookrequest_list
 from domain.services.loan_service import service_read_loans_by_user_id
 from domain.services.user_service import service_read_user, service_update_user
 from routes.request.user_request import RouteReqPutUser
+from routes.response.book_review_response import RouteResGetReviewList
 from routes.response.bookrequest_response import RouteResBookRequest, RouteResBookRequestList
 from routes.response.loan_response import RouteResGetLoanList
 from routes.response.user_response import RouteResGetUser, RouteResPutUser
@@ -111,3 +113,20 @@ async def put_user(
     )
     return response
 
+@router.get(
+    "/my-reviews",
+    response_model=RouteResGetReviewList,
+    status_code=status.HTTP_200_OK,
+    summary="회원의 전체 리뷰 목록 조회",
+)
+async def get_all_user_reviews(
+    db: Session = Depends(get_db),
+    current_user=Depends(get_current_active_user)
+):
+    domain_res = await service_read_reviews_by_user_id(current_user.id, db)
+
+    result = RouteResGetReviewList(
+        data=domain_res,
+        count=len(domain_res)
+    )
+    return result
